requirements.txt
numpy~=1.19.5
torch~=1.8.2+cu111
tensorboardx~=2.4
opencv-python~=4.7.0.72
scipy~=1.7.1
timm~=0.4.12
einops~=0.3.2
pillow~=8.3.2
rawpy~=0.18.0
torchvision~=0.9.2+cu111